feat(SearchBar.tsx) : state lifting
